update version to v2.2.1

+++ project/at_surf_f437/applications/lvgl_music_player/src/system_at32f435_437.c
@@ -152,7 +152,7 @@ void system_core_clock_update(void)
         pllrcsfreq = HEXT_VALUE;
       }
 
-      system_core_clock = (pllrcsfreq * pll_ns) / (pll_ms * pll_fr);
+      system_core_clock = (uint32_t)(((uint64_t)pllrcsfreq * pll_ns) / (pll_ms * pll_fr));
       break;
     default:
       system_core_clock = HICK_VALUE;
